s3_public_access/requirements.txt
boto3==1.42.91
botocore==1.42.91
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
fonttools==4.62.1
Jinja2==3.1.6
jmespath==1.1.0
Markdown==3.10.2
MarkupSafe==3.0.3
pillow==12.2.0
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
tinycss2==1.5.1
tinyhtml5==2.1.0
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
